feat: introduced tabs for a dataset page to organise components better
