+++ .changeset/real-clocks-listen.md
@@ -1,0 +1,5 @@
+---
+"geohub": minor
+---
+
+feat: introduced tabs for a dataset page to organise components better
