+++ sites/geohub/src/lib/config/AppConfig/TabNames.ts
@@ -6,5 +6,8 @@ export enum TabNames {
 	STYLE = 'Style',
 	TRANSFORM = 'Transform',
 	FILTER = 'Filter',
-	INFO = 'Info'
+	INFO = 'Info',
+	LINKS = 'Links',
+	PREVIEW = 'Preview',
+	PERMISSIONS = 'Permissions'
 }
